Add a plugin for Highwire Press metadata
This is a new export plugin `HighwirePress.pm` that is hidden by default
but contains a `dataobj_to_html_header` method which will allow it to be
added to the abstract pages in the form of `<meta>` tags, akin to
`Simple` and `DC`.

Highwire Press is a difficult standard to find details about as there
isn't actually a published standard however it is the recommended
metadata format by 'Google Scholar' so this implements all but one of
the fields they mention at
https://scholar.google.com/intl/en/scholar/inclusion.html#indexing as
well as any extra fields mentioned by 'Zotero'
(https://www.zotero.org/support/dev/exposing_metadata).

The notable exception to implementing every field mentioned by 'Google
Scholar' is 'citation_technical_report_number' as I could not find an
equivalent field, our equivalent of 'technical_report' is 'monograph'
however that doesn't have a 'number' field.

The fields this _does_ implement are:
- citation_abstract
- citation_author
- citation_book_title
- citation_conference_title
- citation_cover_date
- citation_date
- citation_dissertation_institution
- citation_doi
- citation_editor
- citation_firstpage
- citation_isbn
- citation_issn
- citation_issue
- citation_journal_article
- citation_journal_title
- citation_keywords
- citation_language
- citation_lastpage
- citation_online_date
- citation_pdf_url
- citation_pmid
- citation_publication_date
- citation_publisher
- citation_series_title
- citation_technical_report_institution
- citation_title
- citation_volume