Improve the pagerange splitting to match the 'Pagerange' dataobj
This adds a `split_pagerange` function based on the
`EPrints::MetaField::Pagerange::render_single_value` method rather than
using a custom regex just for this, improving consistency with how the
pageranges render in general.

This isn't actually functionally identical to
`Pagerange::render_single_value` as that contains an incorrect fix to
https://github.com/eprints/eprints3.4/issues/60 due to one regex
blocking the only thing the next one was watching for. This instead
implements what `render_single_value` appears to have been attempting to
do, for example it will split 'A-1-A-7' into 'A-1' and 'A-7'.

This also tweaks the date parsing regex to be clearer about how the
month and day values are nested, although this doesn't affect the
functionality due to greedy matching.