Mark the new '<meta>' sections with comments
This makes it a bit clearer when scanning through the rather extensive
`<head>` section which bits are associated with which specification.
This is particularly useful pre-3.4.6 where the `name` and `content` are
semi-randomly ordered (and the same appears to currently be true for
3.5.0-beta-1).

While these tags are absolutely not designed for humans to read it feels
unnecessary to actively avoid human readability so this is a very small
change to help slightly if scanning through the tags (which also makes
it easier to see which tags a particular repository has enabled).