Handle multiple fields in a general way in case they have been changed
If a field has been changed from a single field to a multiple field we
want to show it as multiple items so this adds a `simple_value` function
to both `HighwirePress` and `Prism` to convert these basic 'read a field
from an EPrint' items into data to be push to `@tags`.

Fixes #6.