Add subjects to the metatag keywords
This adds the names of the subjects to the end of the keywords, adding
after a ';' for `HighwirePress` and adding in new `prism.keyword` tags
for PRISM.

This uses `render_description` to get the full subject name (such as
'F1001 Canada (General)') rather than just the code (such as 'F1001').

Fixes #5.